+++ src/rules/IndentationClosingBraces.js
@@ -3,7 +3,7 @@ const rule = {
     scope: "file",
     isCodeNarcRule: false,
     fix: {
-        label: "Fix indentation",
+        label: "Fix indentation closing braces",
         type: "function",
         func: allLines => {
             const newFileLines = [];
@@ -30,7 +30,7 @@ const rule = {
                         j++;
                     }
                     // Set new indentation it on this comment line
-                    if (matchingLineIndent) {
+                    if (matchingLineIndent || matchingLineIndent === 0) {
                         line = (" ".repeat(matchingLineIndent) + line.trimStart()).replace(/\t/g, "");
                     }
                 }