+++ test/lint-format.test.js
@@ -66,47 +66,76 @@ describe('Format with API', function () {
         rimraf.sync(tmpDir);
     }).timeout(100000);
 
-    it('(API:file) should format when if else braces added', async () => {
-        const source = getSourceWithIfElseBracesToFormatBefore()
-        const npmGroovyLintConfig = {
-            source: source,
-            format: true,
-            nolintafter: true,
-            output: 'none',
-            verbose: true
-        };
-        const linter = await new NpmGroovyLint(
-            npmGroovyLintConfig, {
-            jdeployRootPath: 'jdeploy-bundle'
-        }).run();
-
-        assert(linter.lintResult.summary.totalFixedNumber >= 4, `4 Errors have been fixed (${linter.lintResult.summary.totalFixedNumber} returned)`);
-        const result = linter.lintResult.files[0].updatedSource;
-        const expectedResult = getSourceWithIfElseBracesToFormatAfter();
-        const effectiveDiff = getDiff(expectedResult, result, source)
-        assert(linter.status === 0, 'Status is 0');
-        assert(effectiveDiff.length === 0, 'Code has been formatted correctly');
-        checkCodeNarcCallsCounter(2);
-    }).timeout(30000);
+    for (const [key, val] of getSamplesMap()) {
+        it('(API:source) custom check ' + key, async () => {
+            await checkRule(key, val);
+        }).timeout(30000);
+    }
 
 });
 
-function getSourceWithIfElseBracesToFormatBefore() {
-    return `
-  private void doSomething(){
-             if (a == 2)
-                 doSomething();
-}
-`;
+async function checkRule(key, check) {
+    const source = check.before;
+    const npmGroovyLintConfig = {
+        source: source,
+        format: true,
+        nolintafter: true,
+        output: 'none',
+        verbose: true
+    };
+    const linter = await new NpmGroovyLint(
+        npmGroovyLintConfig, {
+        jdeployRootPath: 'jdeploy-bundle'
+    }).run();
+
+    assert(linter.lintResult.summary.totalFixedNumber >= check.totalFixed, `${check.totalFixed} Errors have been fixed (${linter.lintResult.summary.totalFixedNumber} returned)`);
+    const result = linter.lintResult.files[0].updatedSource;
+    const expectedResult = check.after;
+    const effectiveDiff = getDiff(expectedResult, result, source)
+    assert(linter.status === 0, 'Status is 0');
+    assert(effectiveDiff.length === 0, 'Code has been formatted correctly');
+    checkCodeNarcCallsCounter(check.codeNarcCallsCounter);
 }
 
-function getSourceWithIfElseBracesToFormatAfter() {
-    return `
+function getSamplesMap() {
+    return new Map([
+        ['SourceWithIfElseBracesToFormat', {
+            totalFixed: 4,
+            codeNarcCallsCounter: 2,
+            before: `
+private void doSomething(){
+            if (a == 2)
+                doSomething();
+}
+`,
+            after: `
 private void doSomething() {
     if (a == 2) {
         doSomething()
     }
 }
-`;
+`
+        }
+        ],
+        ['OnlyNonCodeNarcFormatRules', {
+            totalFixed: 0,
+            codeNarcCallsCounter: 1,
+            before: `
+    // There is a comment badly aligned here
+if (a == 2) {
+            // And here too
+    x = 1
+    }
+`,
+            after: `
+// There is a comment badly aligned here
+if (a == 2) {
+    // And here too
+    x = 1
 }
+`
+        }
+        ]
+    ]);
 
+}